+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-5038] Links between structures
+
 ### Fixed
 
 - [TD-5279] Fixed some inconsistencies in implementation bulk upload
@@ -27,7 +31,6 @@
 - [TD-5171] Enriches domain's parents on DataStructureVersion show
 - [TD-5245] Allow multiple fields for structure modifiers
 - [TD-4517] Adds `df_label` to Rule filters aggregation
-- [TD-5038] Links between structures
 
 ## [4.53.2] 2022-10-20