+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-5037] Added diff information for draft and pending_aproval structure notes
+  in API response
+
 ### Fixed
 
 - [TD-5279] Fixed some inconsistencies in implementation bulk upload